requirements.txt
langchain
langchain-chroma
langchain-core
langchain-community
langchain-ollama
langgraph
langsmith

chromadb
pymupdf
pypdf
python-docx
openpyxl
pydantic
fastapi
uvicorn
httpx
